fixed most tests

+++ file-manager-module/src/main/java/com/foreach/across/modules/filemanager/services/SpringIntegrationFolderResource.java
@@ -1,14 +1,15 @@
 package com.foreach.across.modules.filemanager.services;
 
+import com.foreach.across.modules.filemanager.business.FileRepositoryResource;
+import com.foreach.across.modules.filemanager.business.FileResource;
 import com.foreach.across.modules.filemanager.business.FolderDescriptor;
 import com.foreach.across.modules.filemanager.business.FolderResource;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.integration.file.remote.RemoteFileTemplate;
-import org.springframework.integration.file.remote.session.Session;
 
-import java.io.IOException;
+import java.util.Collection;
 import java.util.Optional;
 
 @RequiredArgsConstructor
@@ -31,31 +32,48 @@ public abstract class SpringIntegrationFolderResource implements FolderResource
 	@Override
 	@SuppressWarnings("unchecked")
 	public boolean delete( boolean deleteChildren ) {
+		if ( !exists() ) {
+			return false;
+		}
+
+		Collection<FileRepositoryResource> resources = findResources( "*" );
+		if ( !deleteChildren && !resources.isEmpty() ) {
+			return false;
+		}
+		if ( !resources.isEmpty() ) {
+			resources.forEach( frr -> {
+				if ( frr instanceof FolderResource ) {
+					( (FolderResource) frr ).delete( deleteChildren );
+				}
+				if ( frr instanceof FileResource ) {
+					( (FileResource) frr ).delete();
+				}
+			} );
+		}
 		return (boolean) remoteFileTemplate.execute( session -> session.rmdir( getPath() ) );
 	}
 
 	@Override
 	public boolean deleteChildren() {
-		try (Session session = remoteFileTemplate.getSession()) {
-			String[] fileNames = session.listNames( getPath() );
-			boolean allRemoved = true;
-			for ( String fileName : fileNames ) {
-				boolean removed = session.remove( fileName );
-				if ( !removed ) {
-					allRemoved = false;
-				}
-			}
-			return allRemoved;
-		}
-		catch ( IOException e ) {
-			e.printStackTrace();
-		}
-		return false;
+		Collection<FileRepositoryResource> resources = findResources( "*" );
+		return !resources.isEmpty() && resources.stream()
+		                                        .allMatch( f -> {
+			                                        if ( f instanceof FileResource ) {
+				                                        return ( (FileResource) f ).delete();
+			                                        }
+			                                        if ( f instanceof FolderResource ) {
+				                                        return ( (FolderResource) f ).delete( true );
+			                                        }
+			                                        return false;
+		                                        } );
 	}
 
 	@Override
 	@SuppressWarnings("unchecked")
 	public boolean create() {
+		if ( exists() ) {
+			return false;
+		}
 		Optional<FolderResource> optionalParent = getParentFolderResource();
 		if ( optionalParent.isPresent() ) {
 			FolderResource parent = optionalParent.get();
@@ -86,5 +104,10 @@ public abstract class SpringIntegrationFolderResource implements FolderResource
 	@Override
 	public int hashCode() {
 		return folderDescriptor.hashCode();
+	}
+
+	@Override
+	public String toString() {
+		return "axfs [" + folderDescriptor.toString() + "] -> " + String.format( "FTP folder[path='%s']", getPath() );
 	}
 }
